Beta v.2
Final Beta Version

+++ Auto-Admin/freeipa.sh
@@ -13,48 +13,57 @@ ipvalid() {
 ipainstall() {
 read -p "Введите имя домена: " domain
 read -p "Введите имя сервера: " servname
+servname=$(echo $servname | tr [:upper:] [:lower:]) >> /dev/null
 while $true; do
-	read -p "Введите IP-адрес интерфейса: " $ipint
+	read -p "Введите IP-адрес интерфейса: " ipint
 	if ipvalid $ipint; then
 		break
 	else
 		echo "Указан неверный IP-адрес"
 	fi
 done
-read -p -s "Введите пароль учётной записи администратора: " passwd
-read -p "Включить поддержку SMB? [Y/N]" smb
+read -s -p "Введите пароль учётной записи администратора: " passwd
+tput cud1
+read -p "Включить поддержку SMB? [Y/N] " smb
 case $smb in
 	Y | Yes | Да | Д | y | yes | д | да)
-		astra-freeipa-server -d $domain -n $servername -ip $ipint -p $passwd --dogtag -s -y
+		sudo astra-freeipa-server -d $domain -n $servname -ip $ipint -p $passwd --dogtag -s -y
 		;;
 	N | No | n | no | Нет | Н | н | нет)
-		astra-freeipa-server -d $domain -n $servername -ip $ipint -p $passwd --dogtag -y
+		sudo astra-freeipa-server -d $domain -n $servname -ip $ipint -p $passwd --dogtag -y
 		;;
 esac
 }
 
-source /etc/auto-admin/ddns.sh
-echo $doname
-echo $servname
-echo $ipint
+source /etc/auto-admin/.ddnsvars
+
 ping_test="8.8.8.8"
 ping_output=$(ping -c 1 $ping_test)
 
 if echo $ping_output | grep -q "1 packets transmitted, 1 received"; then
-	sudo apt install astra-freeipa-server -y
+	sleep 1
+	if command -v astra-freeipa-server >> /dev/null
+	then
+		sleep 0.001
+	else
+		sleep 0.5
+		sudo apt install astra-freeipa-server -y
+	fi
 	if [[ "$doname" != "" ]]; then
-		read -p "На сервере уже установлен DDNS, хотите использовать его параментры? [Y/N]" answ
+		read -p "На сервере уже установлен DDNS, хотите использовать его параментры? [Y/N] " answ
 		case $answ in
 			Y | Yes | Да | Д | y | yes | д | да)
-				read -p -s "Введите пароль учётной записи администратора: " passwd
-				read -p "Включить поддержку SMB? [Y/N]" smb
+				servname=$(echo $servname | tr [:upper:] [:lower:]) >> /dev/null
+				read -s -p "Введите пароль учётной записи администратора: " passwd
+				tput cud1
+				read -p "Включить поддержку SMB? [Y/N] " smb
 				case $smb in
 				       Y | Yes | Да | Д | y | yes | д | да)
-        			         astra-freeipa-server -d $doname -n $servname -ip $ipint -p $passwd --dogtag -s -y
-			                 ;;
+        			        sudo astra-freeipa-server -d $doname -n $servname -ip $ipint -p $passwd --dogtag -s -y
+			                ;;
  				       N | No | n | no | Нет | Н | н | нет)
-  			        	 astra-freeipa-server -d $doname -n $servname -ip $ipint -p $passwd --dogtag -y
-        			         ;;
+  			        	 	sudo astra-freeipa-server -d $doname -n $servname -ip $ipint -p $passwd --dogtag -y
+        			        ;;
 				esac
 				;;
 			N | No | n | no | Нет | Н | н | нет)
@@ -62,6 +71,7 @@ if echo $ping_output | grep -q "1 packets transmitted, 1 received"; then
 				;;
 		esac
 	else
+		echo ""
 		ipainstall
 	fi
 else
